+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Unreleased
 - Update the cookbook dependency pinned versions
 - Sign out of Mas before trying to sign in as another user
 - Remove all system_user attributes and properties
+- Respect a version property in mac_app_store_mas::install
 
 v2.1.0 (2016-06-08)
 -------------------